fix(cli): fix templates & move some utils to base generator
