+++ packages/cli/generators/controller/templates/src/controllers/controller-rest-template.ts.ejs
@@ -7,7 +7,7 @@ import {
   patch,
   del,
   requestBody
-} from '@loopback/openapi-v3';
+} from '@loopback/rest';
 import {<%= modelName %>} from '../models';
 import {<%= repositoryName %>} from '../repositories';
