+++ packages/cli/generators/controller/templates/src/controllers/controller-template.ts.ejs
@@ -1,6 +1,6 @@
 // Uncomment these imports to begin using these cool features!
 
-// import {inject} from @loopback/context;
+// import {inject} from '@loopback/context';
 
 
 export class <%= name %>Controller {
