+++ packages/cli/lib/base-generator.js
@@ -4,9 +4,12 @@
 // License text available at https://opensource.org/licenses/MIT
 
 'use strict';
+
 const Generator = require('yeoman-generator');
 const chalk = require('chalk');
-
+const debug = require('./debug')('artifact-generator');
+const utils = require('./utils');
+const StatusConflicter = utils.StatusConflicter;
 /**
  * Base Generator for LoopBack 4
  */
@@ -14,6 +17,10 @@ module.exports = class BaseGenerator extends Generator {
   // Note: arguments and options should be defined in the constructor.
   constructor(args, opts) {
     super(args, opts);
+    this.conflicter = new StatusConflicter(
+      this.env.adapter,
+      this.options.force,
+    );
     this._setupGenerator();
   }
 
@@ -21,7 +28,7 @@ module.exports = class BaseGenerator extends Generator {
    * Subclasses can extend _setupGenerator() to set up the generator
    */
   _setupGenerator() {
-    this.artifactInfo = {
+    this.artifactInfo = this.artifactInfo || {
       rootDir: 'src',
     };
   }
@@ -44,6 +51,37 @@ module.exports = class BaseGenerator extends Generator {
     // exit(), exit(undefined), exit('') should exit
     if (!reason) reason = true;
     this.exitGeneration = reason;
+  }
+
+  /**
+   * Checks if current directory is a LoopBack project by checking for
+   * keyword 'loopback' under 'keywords' attribute in package.json.
+   * 'keywords' is an array
+   */
+  checkLoopBackProject() {
+    debug('Checking for loopback project');
+    if (this.shouldExit()) return false;
+    const pkg = this.fs.readJSON(this.destinationPath('package.json'));
+    const key = 'loopback';
+    if (!pkg) {
+      const err = new Error(
+        'No package.json found in ' +
+          this.destinationRoot() +
+          '. ' +
+          'The command must be run in a LoopBack project.',
+      );
+      this.exit(err);
+      return;
+    }
+    if (!pkg.keywords || !pkg.keywords.includes(key)) {
+      const err = new Error(
+        'No `loopback` keyword found in ' +
+          this.destinationPath('package.json') +
+          '. ' +
+          'The command must be run in a LoopBack project.',
+      );
+      this.exit(err);
+    }
   }
 
   /**
